Unify status API handlers with updates
